Add DAX generator allowing for workflow imports
At the moment creating Pegasus abstract workflow (DAX) representing a
scientific pipeline is tightly coupled with the pipeline itself.  This
is not an optimal situation as changing the workflow management system
(WMS) to other than Pegasus would require rewriting the pipeline.

Introduction of the DAX generator is an attempt to separate these two
concerns: writing a pipeline and writing its representation recognized
by the WMS at hand.  Its role is to convert the pipeline, expressed in
some relatively generic format, to the format understandable by WMS.

In current implementation, the DAX generator recognizes the following
formats:

- GraphML,
- GEXF,
- JSON node-link.